move mistral functions to chat_models.bedrock
